+++ api/main.py
@@ -11,9 +11,8 @@ from resources.files import files_router
 from resources.product import product_router
 from resources.promotion import promo_router
 from util.about import config, description
-from util.crypto import encrypt_message, decrypt_message, APP_KEY
 # migration
-# from dao.sql.sql_alchemy_models import init
+from dao.sql_alchemy_models import init
 
 
 # Create the APP
@@ -25,7 +24,7 @@ app = FastAPI(
 )
 
 # migration - create tables
-# init()
+init()
 
 # logging config
 logging.basicConfig(